pyproject.toml
[project]
name = "timetable-viewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "python-dotenv>=1.0.1",
    "supabase>=2.13.0",
]
